requirements.txt
selenium
webdriver-manager
scikit-learn
langchain
ollama
mysql-connector-python
streamlit